Move surface shortcut regression to focused test
